pyproject.toml
[project]
name = "realtime-sector-forecasting"
version = "0.1.0"
description = "Real-time multivariate stock forecasting using PatchTST and Alpaca WebSockets."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alpaca-py>=0.30.0",        # Modern Alpaca SDK
    "neuralforecast>=1.7.0",    # Official Nixtla implementation of PatchTST
    "pandas>=2.0.0",            # Data manipulation
    "numpy<2.0.0",              # Numerical computing (pinned for compatibility)
    "python-dotenv>=1.0.0",     # Secret management for API keys
    "torch>=2.0.0",             # Backend for PatchTST/Timer
    "scikit-learn>=1.3.0",      # Preprocessing and scaling
    "sqlalchemy>=2.0.0",        # Database ORM
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel",                # For Jupyter testing
    "matplotlib",               # Visualization
    "plotly",                   # Interactive stock charts
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend", "model"]